+++ wemake_python_styleguide/presets/types/file_tokens.py
@@ -22,6 +22,7 @@ PRESET: Final = (
 
     statements.ExtraIndentationVisitor,
     statements.BracketLocationVisitor,
+    statements.MultilineStringVisitor,
 
     conditions.IfElseVisitor,
 )